nutanix: allow creating manifests without Prism Central connection
In Assisted Installer we create manifests and generate Ignition in the cloud and pass ignition files to the agents. For platform integration we can't request credentials from users, instead we fill these in with fake credentials and let them change these later.

This commit moves validations to ValidateForProvisioning, so that `openshift-install create manifests` phase would pass with fake credentials