Don't extract entire machine-os-images image
The --file option extracts to the current directory. The --path option
doesn't just set the location to extract to, it says what to extract
where (in scp-style format). By specifying / as the source, we end up
extracting the entire container filesystem when we only want one file.

This reduces the time required to extract the ISO by almost 3 minutes,
to around 15s.