Pull release image before using it
The release-image.sh script uses "podman inspect" to get the digest of
the image, which fails if the image has not already been pulled.
Although it still works OK (using the original pullspec), this results
in a misleading warning in the journal.

To avoid this, pull the release payload before looking up any images in
it.