Write out template from string instead of marshaling from struct
NetworkConfig.Raw does not marshal correctly and the embedded
yaml string writes out as null. So the template has been switched
to a string.

Change AgentConfig to use the PersistToFile interface so that
it observes the --dir command line option.

Validation is turned off for the template and its contents is not
included in asset graph.