OpenStack: Store ignition file in Swift when the upload in Glance fail
Some openstack environments do not allow users to upload or download image data.
This is because the cloud providers prevent them from using illegal images on
the openstack or getting proprietary images on a local.

This PR is based on the Swift functions removed in PR #2960 and adds the feature
to store a ignition file in Swift when an upload in Glance fails, as follows:

   Glance | Swift  | used
   enable | enable | storage
  --------+--------+---------
      v   |   v    | Glance
      v   |        | Glance
          |   v    | Swift   <- this PR
          |        | none

Signed-off-by: ERAMOTO Masaya <eramoto.masaya@jp.fujitsu.com>