Collapse AgentConfig spec
The InstallConfig schema does not have a Spec subresource, so
configuration items are added directly at the top level of the file. It
is desirable that the AgentConfig should behave the same. There is no
Status subresource for an AgentConfig either, and in fact not even any
plans to ever install AgentConfig into clusters as a CRD, so there is no
need for a Spec.

This is a breaking change that will require users to change their
existing agent-config.yaml files, so better that it happen now.