AGENT-246: Temporarily require at least one NMStateConfig to be specified
Numerous places assume that there will be at least one NMStateConfig:

* https://bugzilla.redhat.com/show_bug.cgi?id=2116489
* https://bugzilla.redhat.com/show_bug.cgi?id=2115770
* https://bugzilla.redhat.com/show_bug.cgi?id=2115798
* https://bugzilla.redhat.com/show_bug.cgi?id=2115803
* https://bugzilla.redhat.com/show_bug.cgi?id=2117302

Since these are not easily discoverable by users, temporarily reinstate
the requirement that at least one NMState configuration be provided
(even if the RendezvousIP is specified). This will provide users with an
upfront error rather than a difficult-to-debug mystery.