Base generated NMStateConfig on InstallConfig name
We don't want to require that the user provided the same (or, indeed,
any) name for the AgentConfig as for InstallConfig. So if NMStateConfig
is generated from the AgentConfig, use the cluster name from the
InstallConfig to do it.

Ideally if the user provided ZTP manifests *and* an AgentConfig, we
could generate just the NMStateConfig asset from the AgentConfig and
have the name match the one that appears in the InfraEnv. However, that
is not possible because it would entail a dependency between two
resources (NMStateConfig and InfraEnv) generated by the same command
(agent create cluster-manifests).

If no InstallConfig is provided, we will fall back to the default
cluster name defined in OptionalInstallConfig.