pkg/asset/ignition/bootstrap/cvoignore: Add group/kind/name(space) collision detection
docs/design/resource_dep.svg shows the complicated hierarchy of
installer assets, and it is likely that not all users are familiar
with all the manifests that the installer generates.  It's also
possible that there is a complicated hierarchy on the caller's side,
and not just a single actor, creating additional manifests to feed
into a create-manifests flow.  With all of these inputs, it seems
possible that folks occasionally call the installer with directory
contents that result in multiple manifests for a single resource, as
determined by the (group, kind, namespace, name) tuples.  This commit
adds a check at the slightly-before-Ignition-serialization overrides
asset, which was already in the business of iterating over the content
and building ClusterVersion spec.overrides.  Failing fast before we
build Ignition configs should allow for relatively cheap recovery,
vs. a user's cluster failing to come up, or coming up with a config
that diverges from their intention because some other manifest ended
up taking precedence over the resource that they'd been trying to
control.

Not covered in this commit is deduping between manifests generated on
the bootstrap machine (e.g. by the various 'render' calls to operator
containers).  But that can happen orthogonally, or not, and I think
it's still worth having this cheap, pre-Ignition-config sanity check.