aws: allow clusters with public only subnets
This change allows us to have clusters with public-only subnets without
having to introduce a lot of changes in the installer. Such clusters can
be used in a NAT-less gateway scenario, therefore decreasing costs in cases
where node security is not a concern (e.g, ephemeral clusters in CI).

It's gated through an env var since it's supposed to be only used in CI
for the time being, where an appropriate VPC must be created and passed
to the installer.