gcp: specify network_project_id for network and subnets
This change modifies the GCP Terraform data resources for the
pre-existing networks and subnets to specify the project as defined
in the network_project_id variable. This variable is derived from the
install config networkProjectID parameter. When defined, this causes
terraform to look in the specified project for these resources. When
undefined, the default, this variable will be an empty string at which
point Terraform will look in the default project.