AGENT-325 Set NetworkType in AgentClusterInstall
The IgnitionConfig NetworkType value should be transferred to
the AgentClusterInstall. This will allow the installed cluster
to have the correct network type installed. Currently, the
installed cluster defaults to OpenShiftSDN even if the
NetworkType is set to OVNKubernetes.

If the NetworkType is unspecified, the value is defaulted to
the IgnitionConfig's default value: OpenShiftSDN. A message
is logged to the console indicating that the value was unspecified
and defaulted to OpenShiftSDN.