powervs: use the new name field for the dhcp nw
this will make the destroy logic more robust, as currently we have no way of knowing which networks belonged to which clusters. the current logic is to look at the nw attached to a cluster vm and delete it, but if the cluster vms had been destroyed already, there would be an unused nw left behind.

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>