azure: create bootimages using image gallery
In order to support provisioning of arm64 bootimages, these need to be
created using the image gallery resource as they are not supported
through managed images. For this purpose, it would be good to unify the
implementation of x86 and arm64 to use image galleries as Microsoft have
said that it will the supported way of creating bootimages going forward
for new instance types of new architectures.

This implementation would create one image gallery with two images: a
gen1 and a gen2 image which we would then pick appropriately based on
the VM generation.