Add much longer start timeout
In the ExecStartPre scripts for assisted-service-pod.service, we have to
pull the entire release image. This is fast on a local connection, but
may well not be in real life. Allow much more time for this to run
before restarting the pod, since restarting does not always result in
dependent services coming up successfully.