GCP: create service account for signing URL
Signed URLs, which are used to provide access to the bootstrap
ignition storage object, require a service account key (the private
key is used to sign the url). This works fine when the installer is
authenticated with a service account json file.

In order to allow authentication of the installer with other methods,
this commit creates a service account for the bootstrap process,
and uses that private key to sign the url.