disable packet loss checks
Due to the current behavior of using bond link interfaces for
connectivity checks, the arp tables get screwed up for the bonds where
the link interfaces are teamed up. The fix for that is comming on
OCP 4.12 version of assisted installation. In the meantime, we'll
disable the connectivity check so bonds continue working.

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>