aws: vendor related changes after overriding openshift-api
The openshift/api was overrided to use an api field called lbType in the installer.

This commit contains vendor directory changes which were made after go mod edit the openshift/api
followed by updating of k8s.io/apimachinery, k8s.io/api and k8s.io/client-go to v0.25.2.

This was done so that the installer had all the necessary packages, dependencies and the required code to be used.