aws user tags: set limit == openshift/api
openshift/api defines a limit for user tags, and this sets our
validation to be equal to that. It would be nice if we could vendor
the limit, but the limit is set within kubebuilder markers, which
are not easily digestable in the installer. Therefore, let's just
set a constant value with a reference.

Also cleans up the validation code and adds a test.