gcp: Fix text cases for GCP quota check.
There is a customer case where the customer could not pass the
GCP quota check with only the required services enabled and
looked into the test cases. This one in particular seems to
just check if the test fails or not and was not checking the
actual error message, resulting in the test case with a wrong
error message expected to pass.

Changing the error expected field from string to bool to just
check if there is an error when the required services are
not provided.

Support case: https://access.redhat.com/support/cases/#/case/03200212