Always generate a kubeadmin-password file
If we are supplied with the hash of the kubeadmin password then we don't
know the actual password. Nevertheless, when assisted-service calls the
installer, it always expects this file to be generated. Rather than add
special handling for the agent installer, just generate an empty file if
the hash is supplied.