OCPBUGS-2338: Don't use error messages as default values
If we fail to look up the API VIP in DNS, there is nowhere to return the
error. Previously we were setting the error message as
the default value, which resulted in a pretty horrible-looking error:

[platform.baremetal.apiVIPs: Invalid value: []string{"DNS lookup failure: lookup api.test-cluster.test-domain on 10.0.80.11:53: no such host"}: ip <nil> is invalid, platform.baremetal.apiVIPs: Invalid value: "DNS lookup failure: lookup api.test-cluster.test-domain on 10.0.80.11:53: no such host": "DNS lookup failure: lookup api.test-cluster.test-domain on 10.0.80.11:53: no such host" is not a valid IP, platform.baremetal.apiVIPs: Invalid value: "DNS lookup failure: lookup api.test-cluster.test-domain on 10.0.80.11:53: no such host": IP expected to be in one of the machine networks: 192.168.122.0/23]

This is confusing to the user, because they did not in fact provide an
invalid value. It was made particularly egregious by the change from a
single VIP to a list, which means we see the golang []string{} in the
value output.

If we can't automagically detect the correct default, just leave the
field blank and let the error be reported as a missing value.