Fix ingress config with empty lbType on AWS
When generating the ingresses.config.openshift.io/cluster manifest on AWS,
don't set spec.loadBalancer unless the install-config lbType is nonempty.

Before this commit, if lbType was empty, the installer generated a manifest
with the following:

  loadBalancer:
    platform:
      aws: {}
      type: AWS

Whereas loadBalancer, platform, and aws are optional fields, the type
subfield of the aws field is required.  As a consequence, the cluster
bootstrap would fail with the following error:

    "cluster-ingress-02-config.yml": failed to create ingresses.v1.config.openshift.io/cluster -n : Ingress.config.openshift.io "cluster" is invalid: spec.loadBalancer.platform.aws.type: Required value

This commit ensures that the installer doesn't generate an invalid ingress
manifest when lbType is empty.

Follow-up to commit 5d12adcf3076127163301706e0b2479a822227c5.

This commit fixes bug OCPBUGS-2436.

https://issues.redhat.com/browse/OCPBUGS-2436

* pkg/asset/manifests/ingress.go (generateClusterConfig): Don't set
spec.loadBalancer unless we have a nonempty lbType value to put in
spec.loadBalancer.platform.aws.type.