agent-config: Add missing AdditionalNTPSources
In Agent based installation ZTP input, one can provide
AdditionalNTPSources in InfraEnv. The sources configured there are
merged by assisted-service with those that are provided by DHCP option
42 (According to RFC 2132) and provided to the openshift-install that
assisted-service runs as MachineConfigs for the different roles defined
for the cluster nodes.

This commit closes the gap we had when installing from install-config +
agent-config.

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>