Reduce duplicate "Still waiting for cluster to initialize" logs
The IPI wait-for install-complete command uses a watch on the
clusterversion object to log the status when it changes. While the agent
wait-for install-complete command uses similar logging code, it gets the
status by polling rather than using a watch. This results in many more
identical debug messages. To emulate the behaviour of IPI, only print a
debug message when the clusterversion object has changed.