GCP: Only passthrough service accounts on XPN
When using the in-tree provider (rather than CCM which is in tech
preview) the kubelet gets credentials from the VM host service account.
In XPN, the kubelet needs perms in both the host and service projects.
Therefore, when runnning in passthrough mode the installer credentials
are passed through to the control-plane hosts.

This commit restricts this behavior to only occur when running in
GCP XPN which is in tech preview. Customers may not expect the installer
host service account to be attached to the control-plane nodes during
a regular install.