Adapt MOTD for rendezvous host vs other hosts
Make the main message specific to the agent installer, and put it inside
motd.d so that it does not get overwritten on the bootstrap node.

Put info about the primary service to watch (assisted-service on node 0;
agent on other hosts) directly in /etc/motd. On node 0, this will be
overwritten to say that bootkube is the primary service once
bootstrapping starts. Writing this data to /etc/motd also has the effect
of overwriting the stuff about using Machine Config Operator to
configure the host, which CoreOS automatically puts in there but which
doesn't make sense until the cluster is booted.