Redact any pull secrets from the journal in agent-gather
Although we should never log the value of a pull-secret, to be on the
safe side redact any of the pull secret auth values that we know about
from the journal before saving it in the agent-gather archive.

The exported journal format is binary[1], so we must ensure that the
redacted output has the same length as the input.

[1] https://systemd.io/JOURNAL_EXPORT_FORMATS/#journal-export-format