OCPBUGS-3524: data: azurerm: restore RHCOS SA access configuration
In the transition from v2 to v3 of the azurerm terraform provider, not
only was an option renamed from `allow_blob_public_access` to
`allow_nested_items_to_be_public` but its default value was also changed
from `false` to `true` [1].

So to restore the previous access setting for the RHCOS Storage Account,
we need to explicitly set the new option to `false`.

Thanks Jinyun Ma for pointing this out.

[1] https://github.com/hashicorp/terraform-provider-azurerm/blob/main/CHANGELOG.md#300-march-24-2022