WIP Update installer-artifacts to copy binaries
This is one potential solution, but probably not the ideal one. This
is the easiest solution to demonstrate the PoC.

We just need the tf binaries somewhere in the release image. This
solution would still incur an unnecessary cost at every installer
build. It would make sense to move these providers to a separate
image so they would only need to be built (or copied) when THEY
or one of them changes.