vSphere Zonal: Validate DNS and LB
The PR:
https://github.com/openshift/installer/pull/5798
Introduced the ability to install with
the VIPs unset.

This PR adds validation to the vSphere
zonal installation for DNS and LB
existence.

It will query DNS for api and api-int.
If either fail a error will be provided.

It will perform a TCP/6443 port connection
to the `api.ClusterDomain()`. If not
available will emit a warning only.