Bug OCPBUGS-3921: OpenStack: fix bootstrap destroy cmd
The installer bootstrap destroy command does not locate the
clouds.yaml file containing the credentials when it's present
in the current directory. This commit fixes the issue by ensuring
that prior to destroying the resources, the installer attempts to
look for the OS_CLIENT_CONFIG_FILE env var and if unset looks for
any clouds.yaml file in the current directory.