OCPBUGS-4125: Enable CVO unmanage overrides in bootstrap-in-place installations
# Background

#5258 introduced a mechanism to tell CVO to not manage some cluster resources,
as that management was creating races between bootkube and CVO.

In #5271, this mechanism was removed from bootstrap-in-place
installations.

The reason for the removal was because attempting to apply
`original_cvo_overrides.patch` after the cluster-bootstrap podman
container was done wasn't working, as in BIP by the time that that
container is down, there's no longer a running kube-apiserver apply the
patch on (unlike regular installations with a bootstrap node where the
API server keeps running on the actual nodes). The justification for
simply disabling the mechanism instead of finding a solution was based
on an incorrect assumption that "CVO doesn't even run in BIP pre-pivot
so there's no need to perform overrides anyway as there's no CVO to race
with".

# Issue

Naturally, since CVO *is* actually running in the pre-pivot phase of
bootstrap-in-place, not applying CVO overrides to disable the CVO
management of bootkube-created resources led to the same races #5258 was
aiming to solve in bootstrap-in-place installations.

One manifestation of that is OCPBUGS-4125.

# Solution

This commit re-introduces the mechanism that was originally introduced in #5258
(and removed in #5271).

To solve the issue that motivated #5271, this commit doesn't attempt to
apply the `original_cvo_overrides.patch` in bootstrap-in-place
pre-reboot, and instead will transfer `original_cvo_overrides.patch` through
`master-update.fcc` to the post-pivot phase and the bootstrap-in-place
post-reboot script will make sure to apply the patch after the API
server is running again.