OCPBUGS-4367: Fix missing debug messages when getting baseISO
A new logger is currently created for the baseISO and oc commands.
Since the default log level is not debug, this results in the messages
not being output. It is cleaner to use the global logger which prints
the debug log messages by default.

In addition, a log message was added for a failure running the oc
command due to the --icspFile option as this indicates that an older
oc binary is being used which doesn't support this option.