OCPBUGS-3987: Check nmstateconfig content in agent-config.yaml
The NMStateConfig data in agent-config.yaml was not being validated,
it was only checked in the Load() i.e. the data in nmstateconfig.yaml.
This can result in invalid configurations being stored in ignition
and then the install would fail when these configs were validated
by assisted-service.

The change is to do the same checks in Load() and Generate().