azure: extract ODataError for more friendly message
Errors returned by the new Azure SDK are not very helpful. So before
returning any error back to the user, we try to extract a more friendly
"code: message" format from them.

Please note that we do that at the very end so that unwraping and
checking for other non-ODataError types is still possible during the
function call chain.

https://github.com/microsoftgraph/msgraph-sdk-go/discussions/218