OpenStack: Force JSON content-type in Swift object listing
Some OpenStack object storages respond with `204 No Content` to list
requests when there are no containers or objects to list. In these
cases, when responding to requests with an `Accept: text/plain` or no
`Accept` header, some object storages omit the `content-type` header in
their status-204 responses.

Now, Gophercloud throws an error when the response does not contain a
`content-type` header.

With this change, we work around the issue by forcing Gophercloud to
request a JSON response from the object storage when listing objects.
When passed an `Accept: application/json` header, the server responds
with `200 Ok` and a `content-type` header in our tests.

This solution gives us a fix that is easily backportable because it
doesn't require any dependency bump.