[release-4.12] OCPBUGS-4995: Azure: toggle image in machinesets (#6715)
* Azure: toggle image in machinesets

Hive vendors the installer and uses the asset package to generate
machinesets for scaleup. Because Hive is using the latest code version
but installing multiple previous versions, the machinesets--particularly
the values--need to be backward compatible.

In this particular case, the installer switched from using Azure
managed images to image galleries in 4.12. In 4.12+ Azure machinesets
expect an image referencing an image gallery, while prior to this change
the machinesets looked for a managed image.

This commit updates the machineset code to allow a toggle which will
allow Hive to generate Azure machinesets utilizing managed images,
which should be done with 4.11 and earlier clusters.

This change also future proofs the 4.12+ by switching the machinesets
to use the latest version, rather than tying them to a particular RHCOS
version.

* fixup! hyperv gen support for managed image

Co-authored-by: patrickdillon <padillon@redhat.com>