Azure MachineSets: Expose useImageGallery to consumers
This allows end-users of the MachineSets() function to choose whether
they want to use the image gallery or not (<=4.11 vs >=4.12).
Additionally, the unused parameter rhcosVerison has been removed, and
a variable has been shadowed to appease the golint CI check.

See #6694 for more details concerning why this change is necessary.