AGENT-501 Make it easier to change node zero IP and SERVICE_BASE_URL
The node zero IP and ServiceBaseURL are hard-coded into scripts and
systemd service definition files. This makes it difficult to change
the node zero IP once the agent ISO has booted up. To faciliate
changing these settings in the interactive network configuration
use case, the hard-codings have been replaced with sourcing files
containing environment variables.

In the interactive network configuration scenario, the environment
files will be updated by the network configuration utility. The
agent services are blocked by the interactive console service. When
the interactive console service finishes, the agent services start
up and read in the environment files picking up any changes.

In the non-interactive scenario, the environment files will be
prepopulated with values for node zero IP and SERVICE_BASE_URL in
the Agent ISO. This is how it works today.

Signed-off-by: Richard Su <rwsu@redhat.com>