AGENT-453: Create interactive console service for agent installer
Users would like the ability to change their network
configuration at the console if network connectivity problems
are detected.

To achieve this goal, this patch adds a new service called
agent-interactive-console.service to block the login prompt
and the agent services that pulls an image from the registry.

The service will execute the agent TUI to allow users to update
their network configuration. The TUI will check there is
connectivity to the registry and to the rendezvous host. If
the connectivity checks pass, the TUI exits, which also lead
the interactive console service to exit, and this unblocks the
login prompt and agent services waiting for pull from the registry,
allowing the agent-based installer to proceed.

The agent TUI will be added in a future patch.

For now, the service executes a script that logs its
presence, sleeps for 60 seconds, and exits. This should not
block the automated flow.

Added ConditionPathExists=/usr/local/bin/agent-tui, which means
the service does not start nor is it active until the agent-tui
binary is present at that path.

Most of the service definition was lifted from celebdor's POC:
https://github.com/openshift/installer/pull/6560/

Signed-off-by: Richard Su <rwsu@redhat.com>