openstack: validate dns and external load balancer
This PR adds validation to the OpenStack installation for DNS and LB
existence.

It will query DNS for api and api-int. If either fail a error will be provided.

It will perform a TCP/6443 port connection to the api.ClusterDomain(). If not
available will emit a warning only.

We are re-using the function from vsphere/validation (we had to make it
public).