PowerVS: Check for potential network conflict in existing CloudConnections
When we check the DHCP subnets now, we just check for their existence, because it
was only possible to create a network with 192.168.X.0/24.  Now that we can specify
a subnet and share a cloud connection, we need to make the quota check more granular.
IOW, it should only fail if there is an existing DHCP network with the same subnet
we're going to create.

https://issues.redhat.com/browse/MULTIARCH-2752