CORS-2496: Default `2xlarge` instead of `xlarge` in AWS when the control-plane is single-node
# Background

* Starting with version 4.13 OCP is going to officially support Single
  Node clusters on AWS.

* The minimum documented OCP requirement for single-node control plane
  nodes is 8-cores and 16GiB of RAM

* The current default instance type chosen for AWS clusters by the
  installer is `xlarge` which is 4 cores and 16GiB of RAM

# Issue

The default instance type the installer currently chooses for Single
Node Openshift clusters doesn't follow our documented minimum
requirements

# Solution

When the number of replicas of the ControlPlane pool is 1, the installer
will now choose `2xlarge` instead of `xlarge`.

# Caveat

`2xlarge` has 32GiB of RAM, which is twice as much as we need, but it's
the best we can do to meet the minimum single-node requirements, because
AWS doesn't offer a 16GiB RAM instance type with 8 cores.