OCPBUGS-6725: In wait-for, use consistent delay when waiting for APIs
If the Kube API takes a while to come up we can end up with the message
'INFO cluster bootstrap did not complete'
repeated every 2 seconds as shown in the bug.
Treat both APIs consistently so that there is a 30 second delay
before repeating the message, and the output is more descriptive.