OCPBUGS-6919: vsphere: fix validation of CPUS and CoresPerSocket (#6822)
* OCPBUGS-6919: vsphere: fix validation of CPUS and CoresPerSocket

I think the original intention behind the validation was to check that,
in case `CoresPerSocket` was not set by the user, it would be checked
against a default value (currently set to 4). However, that's not what
is happening in practice, and both the check against the user-inputed
value and the default were being performed.

That means that using CPUS=2 and CoresPerSocket=2 as exemplified in the
official docs [1] would result in a validation error:
```
failed to create install config: invalid "install-config.yaml" file: controlPlane.platform.vsphere.cpus:
Invalid value: 2: numCPUs specified should be a multiple of cores per socket which is by default 4]
```
even though it's a valid configuration (albeit a pretty tiny compute
node).

I'm changing the code a bit to make it more clear what the defaults
values are representing and fixing the checks so they are made only once
with the correct values for CPUs and CoresPerSocket.

[1] https://docs.openshift.com/container-platform/4.11/installing/installing_vsphere/installing-vsphere-installer-provisioned-customizations.html#installation-installer-provisioned-vsphere-config-yaml_installing-vsphere-installer-provisioned-customizations

* fixup! OCPBUGS-6919: vsphere: fix validation of CPUS and CoresPerSocket

* fixup! OCPBUGS-6919: vsphere: fix validation of CPUS and CoresPerSocket