RFE-2782/phase-1: Set MTU to CNI when edge pool is present on AWS
This change generates the CNO manifest file
`cluster-network-03-config.yml` into the cluster manifest path.

The goal is to force the Cluster Network to use MTU according to the
network plugin when the AWS Edge machine pool is set. Currently,
there is support only for AWS Local Zones on the edge machine pool.
Local Zones limit the MTU to 1300[1] to communicate with the
nodes in Availability Zones (regular zones within the region).

In Phase-0[2] of support of AWS Local Zones, the instructions are added
in the documentation[3]. This change is covering Phase-1[4].

[1] https://docs.aws.amazon.com/local-zones/latest/ug/how-local-zones-work.html
[2] https://issues.redhat.com/browse/SPLAT-635
[3] https://github.com/openshift/openshift-docs/pull/55327
[4] https://issues.redhat.com/browse/SPLAT-636