vSphere: Platform spec deprecation
To make vSphere installation easier for zonal
and provide additional options for non-zonal
installation the existing fields will be
depreciated with this commit.

To facilitate this we extended the networking
converstion logic to be specific for vsphere.
The conversion will take existing IPI, UPI
and 4.12 Zonal platform specs and modify
to work with the new vcenters and failure domains.

Additional Details:
- Updated agent installer's use of platform spec fields
- Add additional validation for vSphere versions of
vCenter and ESXi hosts within requested clusters.
- Modify validations to failure domain and vcenters
requirements - including requiring paths.
- Create pkg/types/vsphere/conversion/installconfig.go
to modify a given platform spec for the new requirements
- Modified platform spec renaming fields depreciated
as required.
- DNS and LB check broken apart to assist with unit testing.
- Changes from linting.